+++ lib/praxis/tasks/routes.rb
@@ -2,11 +2,16 @@ namespace :praxis do
 
   desc 'List routes, format=json or table, default table'
   task :routes, [:format] => [:environment] do |t, args|
-    require 'ruport'
+    require 'terminal-table'
 
-    table = Table(:version, :path, :verb, :resource, 
-      :action, :implementation, :name, :primary)
 
+    table = Terminal::Table.new title: "Routes",
+    headings:  [
+      "Version", "Path", "Verb",
+      "Resource", "Action", "Implementation", "Name", "Primary"
+    ]
+
+    rows = []
     Praxis::Application.instance.resource_definitions.each do |resource_definition|
       resource_definition.actions.each do |name, action|
         method = begin
@@ -14,11 +19,11 @@ namespace :praxis do
         rescue
           nil
         end
-        
+
         method_name = method ? "#{method.owner.name}##{method.name}" : 'n/a'
 
         action.routes.each do |route|
-          table << {
+          rows << {
             resource: resource_definition.name,
             version: route.version,
             verb: route.verb,
@@ -32,11 +37,18 @@ namespace :praxis do
       end
     end
 
+
+
+
     case args[:format] || "table"
     when "json"
-      puts JSON.pretty_generate(table.collect { |r| r.to_hash })
+      puts JSON.pretty_generate(rows)
     when "table"
-      puts table.to_s
+      rows.each do |row|
+        table.add_row(row.values_at(:version, :path, :verb, :resource,
+                                    :action, :implementation, :name, :primary))
+      end
+      puts table
     else
       raise "unknown output format: #{args[:format]}"
     end